addons-report/report_extend_bf/requirements.txt
py3o.template==0.9.12
genshi==0.8.dev0
